+++ src/extend-context-menu.ts
@@ -1,8 +1,8 @@
 // @ts-ignore  TODO: remove tag when onContextMenu is declared in nebula/stardust
 import { onContextMenu } from '@nebula.js/stardust';
-import copyCellValue from '../table/utils/copy-utils';
+import copyCellValue from './table/utils/copy-utils';
 
-export default function useContextMenu() {
+export default function extendContextMenu() {
   onContextMenu?.((menu: any, event: any) => {
     event.target &&
       event.target.closest('.sn-table-cell') &&
