+++ src/table/pagination-table/components/TableWrapper.tsx
@@ -42,14 +42,40 @@ function TableWrapper(props: TableWrapperProps) {
   const tableWrapperRef = useRef<HTMLDivElement>(null);
 
   const { yScrollbarWidth } = useScrollbarWidth(tableContainerRef);
-
-  const setShouldRefocus = useCallback(() => {
-    shouldRefocus.current = rootElement.getElementsByTagName('table')[0].contains(document.activeElement);
-  }, [rootElement]);
+  useFocusListener(tableWrapperRef, shouldRefocus, keyboard);
+  useScrollListener(tableContainerRef, direction);
+  useKeyboardActiveListener();
 
   useEffect(() => {
     tableContainerRef.current?.scrollTo(0, 0);
   }, [pageInfo, totalRowCount]);
+
+  // Except for first render, whenever the size of the data (number of rows per page, rows, columns) or page changes,
+  // reset tabindex to first cell. If some cell had focus, focus the first cell as well.
+  useDidUpdateEffect(() => {
+    resetFocus({
+      focusedCellCoord,
+      rootElement,
+      shouldRefocus,
+      setFocusedCellCoord,
+      isSelectionMode,
+      keyboard,
+      announce,
+      totalsPosition,
+    });
+  }, [rows.length, totalRowCount, totalColumnCount, page]);
+
+  useDidUpdateEffect(() => {
+    setYScrollbarWidth(yScrollbarWidth);
+  }, [yScrollbarWidth]);
+
+  useDidUpdateEffect(() => {
+    selectionDispatch({ type: SelectionActions.UPDATE_PAGE_ROWS, payload: { pageRows: rows } });
+  }, [rows]);
+
+  const setShouldRefocus = useCallback(() => {
+    shouldRefocus.current = rootElement.getElementsByTagName('table')[0].contains(document.activeElement);
+  }, [rootElement]);
 
   const handleChangePage = useCallback(
     (pageIdx: number) => {
@@ -74,33 +100,6 @@ function TableWrapper(props: TableWrapperProps) {
       isSelectionMode,
     });
   };
-
-  useFocusListener(tableWrapperRef, shouldRefocus, keyboard);
-  useScrollListener(tableContainerRef, direction);
-  useKeyboardActiveListener();
-
-  // Except for first render, whenever the size of the data (number of rows per page, rows, columns) or page changes,
-  // reset tabindex to first cell. If some cell had focus, focus the first cell as well.
-  useDidUpdateEffect(() => {
-    resetFocus({
-      focusedCellCoord,
-      rootElement,
-      shouldRefocus,
-      setFocusedCellCoord,
-      isSelectionMode,
-      keyboard,
-      announce,
-      totalsPosition,
-    });
-  }, [rows.length, totalRowCount, totalColumnCount, page]);
-
-  useDidUpdateEffect(() => {
-    setYScrollbarWidth(yScrollbarWidth);
-  }, [yScrollbarWidth]);
-
-  useDidUpdateEffect(() => {
-    selectionDispatch({ type: SelectionActions.UPDATE_PAGE_ROWS, payload: { pageRows: rows } });
-  }, [rows]);
 
   const tableAriaLabel = `${translator.get('SNTable.Accessibility.RowsAndColumns', [
     String(rows.length + 1),
