+++ src/table/pagination-table/components/body/withSelections.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useContextSelector, TableContext } from '../../../context';
 import { getSelectionStyle } from '../../../utils/styling-utils';
 import { getCellSelectionState } from '../../../utils/selections-utils';
-import { getSelectionMouseHandlers } from '../../../utils/handle-click';
+import { getSelectionMouseHandlers } from '../../../utils/handle-mouse';
 import { CellHOC, CellHOCProps } from '../../../types';
 
 export default function withSelections(CellComponent: CellHOC) {
